pyproject.toml
[project]
name = "backend-test-task-fastapi"
version = "0.1.0"
description = "Test task backend FastAPI"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "bleach>=6.3.0",
    "fastapi[standard]>=0.129.0",
    "loguru>=0.7.3",
    "passlib[bcrypt]<2.0.0,>=1.7.4",
    "bcrypt==4.0.1",
    "psycopg[binary]>=3.3.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "tenacity>=9.1.4",
    "types-passlib>=1.7.7.20260211",
    "types-python-jose>=3.5.0.20250531",
    "types-bleach>=6.3.0.20251115",
    "types-python-slugify>=8.0.2.20240310",
    "python-slugify>=8.0.4",
    "redis[async]>=7.2.0",
    "aioredis>=2.0.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "aiosqlite>=0.22.1",
    "fakeredis>=2.34.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.4",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
